refactor: split record/mod.rs into focused sub-modules (#942)
Extract the large unit test block (~730 lines) from record/mod.rs into
a dedicated tests.rs file. This reduces mod.rs from 808 lines to 79
lines (well under the 200-line target), following the project convention
for organising large test modules into separate files.

The record/ directory now contains:
- mod.rs (79 lines) — orchestration and re-exports
- validation.rs (84 lines) — input validation
- processing.rs (89 lines) — record building
- tests.rs (729 lines) — unit tests

Public API unchanged, all tests pass unmodified.
